refactor(deck-settings): theme tab-bar active state and nest designer-tabs keys
Switch tab-bar's selected state from base brown/grey to --theme-primary
so it picks up the surrounding deck theme. Relocate the
designer-tabs.* locale keys under deck.settings-modal.design.* so all
design-tab strings group together.
