+++ src/components/modals/deck-settings/tab-design/index.vue
@@ -3,7 +3,7 @@ import { computed, inject } from 'vue'
 import { useI18n } from 'vue-i18n'
 import CoverDesigner from '@/components/deck/cover-designer/index.vue'
 import CardDesigner from './card-designer/index.vue'
-import TabBar from './tab-bar.vue'
+import TabBar from '@/components/layout-kit/tab-bar.vue'
 import DeckPreview from '@/components/deck/deck-preview.vue'
 import { deckEditorKey } from '@/composables/deck-editor'
 import { useMobileBreakpoint } from '@/composables/use-media-query'
@@ -16,9 +16,9 @@ const editor = inject(deckEditorKey)!
 const is_mobile = useMobileBreakpoint('md')
 
 const tabs = computed<SideTab[]>(() => [
-  { value: 'cover', label: t('deck.settings-modal.designer-tabs.cover') },
-  { value: 'front', label: t('deck.settings-modal.designer-tabs.front') },
-  { value: 'back', label: t('deck.settings-modal.designer-tabs.back') }
+  { value: 'cover', label: t('deck.settings-modal.design.designer-tabs.cover') },
+  { value: 'front', label: t('deck.settings-modal.design.designer-tabs.front') },
+  { value: 'back', label: t('deck.settings-modal.design.designer-tabs.back') }
 ])
 
 const card_side_attributes = computed(() =>
