+++ src/locales/en-us.json
@@ -256,9 +256,9 @@
   "deck.settings-modal.danger-zone.reset-button": "Reset Reviews",
   "deck.settings-modal.design.card-designer.text-size-label": "Text Size",
   "deck.settings-modal.design.card-designer.alignment-label": "Alignment",
-  "deck.settings-modal.designer-tabs.cover": "Cover",
-  "deck.settings-modal.designer-tabs.front": "Front",
-  "deck.settings-modal.designer-tabs.back": "Back",
+  "deck.settings-modal.design.designer-tabs.cover": "Cover",
+  "deck.settings-modal.design.designer-tabs.front": "Front",
+  "deck.settings-modal.design.designer-tabs.back": "Back",
   "deck.settings-modal.cover.bg-color": "Theme",
   "deck.settings-modal.cover.bg-color-picker.label": "BG Color",
   "deck.settings-modal.cover.bg-color-picker.border-label": "Border",
